+++ src/drivers/rc_input/RCInput.hpp
@@ -115,7 +115,7 @@ private:
 	void fill_rc_in(uint16_t raw_rc_count_local,
 			uint16_t raw_rc_values_local[input_rc_s::RC_INPUT_MAX_CHANNELS],
 			hrt_abstime now, bool frame_drop, bool failsafe,
-			unsigned frame_drops, int rssi);
+			unsigned frame_drops, uint8_t rssi);
 
 	void set_rc_scan_state(RC_SCAN _rc_scan_state);
